solution 1.1: simple tests with printf - passing

+++ components/calc/src/calc.h
@@ -8,8 +8,8 @@
  *
  * @param a int16_t: First integer
  * @param b int16_t: Second integer
- * @return int16_t: Sum of a and b
+ * @return int32_t: Sum of a and b
  */
-extern int16_t calc_add(int16_t a, int16_t b);
+extern int32_t calc_add(int16_t a, int16_t b);
 
 #endif /* CALC_H */
